Require SSI & friends to pull in postpone
The nginx build system was modified in 1.9.11 in a way that breaks
the hack we had previously been using to pull in the postpone module.

The upstream has expressed that they don't feel this is a flaw in
their build system since ssi is enabled by default (pulling in
postpone).

Rather than pile more hacks on the pile, we're going to throw in
the towel and just document the limitation under the installation
section of the README.

Closes #52
Closes #51
Reference: http://mailman.nginx.org/pipermail/nginx-devel/2016-April/008172.html