Bring back URL decoding and stop marking requests as external
This reverts #96, with the exception of not decoding '+' to ' '.

The change in #96 fixed #95 but introduced #102. It appears to be impossible to fix both bugs without modifying Nginx itself.
However, the second part of the original bug was really a problem with S3 rather than mod_zip, so mod_zip functionality should take priority.

Therefore, bring back original behavior - except for decoding '+' to ' ', which was still a bug in mod_zip.

This fixes #102.