Explicitly define filters chain order, with modules which need to execute before and after mod_zip
Possible fix for #90 when compiled as a dynamic module.

This also eliminates the need to compile addition, slice or ssi when
building this module dynamically.since we're now explicitly defining
the postpone ordering.

Unforutnately, I don't believe the we can explicitly define filter
chain ordering in static modules, it may be possible to fix it by
reordering --add-module in configure, or by editing nginx build files
but that's out of our scope as a plugin.